Building docker image and zero install updates

+++ bootstrap/install.sh
@@ -4,14 +4,15 @@
 # Command Central client tools install script
 ###
 
-# latest public GA version
-CC_VERSION=10.2-fix1
-
-case "`uname`" in
-  Darwin) CC_INSTALLER=cc-def-$CC_VERSION-osx.sh ;;
-   Linux) CC_INSTALLER=cc-def-$CC_VERSION-lnxamd64.sh ;;
-       *) echo "Not supported OS" && exit 1 ;;
-esac
+if [ -z $CC_INSTALLER ]; then
+  # latest public GA version
+  CC_VERSION=${CC_VERSION:-10.2-fix1}
+  case "`uname`" in
+    Darwin) CC_INSTALLER=cc-def-$CC_VERSION-osx.sh ;;
+     Linux) CC_INSTALLER=cc-def-$CC_VERSION-lnxamd64.sh ;;
+         *) echo "Not supported OS" && exit 1 ;;
+  esac
+fi
 
 # default public download site
 URL=${CC_INSTALLER_URL:-http://empowersdc.softwareag.com/ccinstallers}
@@ -22,14 +23,25 @@ export CC_CLI_HOME="$CC_HOME/CommandCentral/client"
 
 mkdir -p "$HOME/Downloads"
 file="$HOME/Downloads/$CC_INSTALLER"
-echo "Downloading ${URL}/${CC_INSTALLER} ..."
-curl -o "$file" "${URL}/${CC_INSTALLER}"
-chmod +x $file
+
+if [ -f "$file" ]; then
+  echo "Found $file"
+else
+  echo "Downloading ${URL}/${CC_INSTALLER} ..."
+  curl -o "$file" "${URL}/${CC_INSTALLER}"
+  chmod +x $file
+fi
+
 echo "Installing ..."
 $file -D CLI -L -d "$CC_HOME"
-# rm $file
 
-# case "`uname`" in
-#     Darwin) ;;
-#      Linux) ln --symbolic "${CC_CLI_HOME}/bin/sagcc" /usr/bin/sagcc && ln --symbolic "${CC_CLI_HOME}/bin/sagccant" /usr/bin/sagccant ;;
-# esac
+echo "Update your environment:"
+
+echo "export CC_CLI_HOME=${CC_CLI_HOME}"
+echo "export ANTCC_HOME=${CC_CLI_HOME}"
+echo "export PATH=$PATH:$CC_CLI_HOME/bin:$ANTCC_HOME/bin"
+
+echo "Verify by running 'antcc help'"
+
+#echo "ln -s ${CC_CLI_HOME}/bin/sagcc /usr/local/bin/"
+#echo "ln -s ${CC_CLI_HOME}/bin/sagccant /usr/local/bin/"
